use a big ugly hack to draw the cursor
To get the cursor into the right positon it has to be moved into the right
position when drawing WindowManager. To do that I really have to figure out the
cursor's position relative to the stdscr in TextInput.

The only way I found to get the right numbers is to recursivly ask all the
parent widgets for their positions and adding them up.

After that WindowManager has to know where to draw the cursor, so it get a
SetCursor methad.