Mark duplicate default key bindings
There should be always one default key for any action. All other keys
that trigger same action should be located in key bindings config. However
there is no support for it yet so it has to be hacked in the code, places
where this happens were marked with a note.