Use pkg-config to find ncursesw
It was previously not possible to use pkg-config to search for ncursesw
because some operating systems (distributions) did not configure ncurses
with --enable-pc-files, but it seems it is now enabled by absolute
majority of them and so it is no longer necessary to use
ncurses[56]-config.

This also allows to get rid of use of the AC_PATH_LIB macro which has been
deprecated for a long time.